Move HasBrooklynManagementContext to o.a.b.core.mgmt
- Was previously in o.a.b.camp.brooklyn.api, in the core module
